Add keystore generation to docker
In order to call other services that use SSL certs generated by our internal self-signed CA, we will need to convert the ca certificate that dr attaches to the container to a jvm trust store. This change adds a command to the container that executes before it starts, and builds the trust store from the ca certificate.